refactor: mv Ledger code into new hardware-wallet domain
