+++ src/__swaps__/screens/Swap/providers/swap-provider.tsx
@@ -43,10 +43,10 @@ import { supportsDelegatedExecution } from '@/features/delegation/willDelegate';
 import { clearCustomGasSettings } from '@/features/gas/hooks/useCustomGas';
 import { getGasSettingsBySpeed, getSelectedGas } from '@/features/gas/hooks/useSelectedGas';
 import { buildGasParams } from '@/features/gas/utils/parseGas';
+import { LedgerSigner } from '@/features/hardware-wallet/utils/LedgerSigner';
 import { backendNetworksActions, useBackendNetworksStore } from '@/features/network/stores/backendNetworksStore';
 import { ChainId } from '@/features/network/types/backendNetworks';
 import { divWorklet, equalWorklet, lessThanOrEqualToWorklet, mulWorklet } from '@/framework/core/safeMath';
-import { LedgerSigner } from '@/handlers/LedgerSigner';
 import { getProvider } from '@/handlers/web3';
 import { WrappedAlert as Alert } from '@/helpers/alert';
 import { type useAnimatedInterval } from '@/hooks/reanimated/useAnimatedInterval';
