+++ src/features/hardware-wallet/components/NanoXDeviceAnimation.tsx
