+++ src/features/rnbw-rewards/utils/claimAirdrop.ts
@@ -2,13 +2,13 @@ import { type Address } from 'viem';
 
 import { analytics } from '@/analytics';
 import { type NativeCurrencyKey } from '@/features/currency/types';
+import { LedgerSigner } from '@/features/hardware-wallet/utils/LedgerSigner';
 import { useAirdropBalanceStore } from '@/features/rnbw-rewards/stores/airdropBalanceStore';
 import { type ClaimAirdropResponse, type ClaimAirdropResult } from '@/features/rnbw-rewards/types/claimAirdropTypes';
 import { getPlatformResult } from '@/features/rnbw-rewards/utils/getPlatformResult';
 import { pollClaimStatus, type PollClaimStatusResult } from '@/features/rnbw-rewards/utils/pollClaimStatus';
 import { time } from '@/framework/core/utils/time';
 import { type RainbowFetchResponse } from '@/framework/data/http/rainbowFetch';
-import { LedgerSigner } from '@/handlers/LedgerSigner';
 import { getProvider } from '@/handlers/web3';
 import { logger, RainbowError } from '@/logger';
 import { loadWallet, signPersonalMessage } from '@/model/wallet';
