+++ src/features/transfer/hooks/useSendSubmit.ts
@@ -18,6 +18,7 @@ import type useENSProfile from '@/features/ens/hooks/useENSProfile';
 import { type ActionTypes } from '@/features/ens/hooks/useENSRegistrationActionHandler';
 import { type REGISTRATION_STEPS } from '@/features/ens/utils/helpers';
 import { parseGasParamsForTransaction } from '@/features/gas/utils/parseGas';
+import { setHardwareTXError } from '@/features/hardware-wallet/navigation/HardwareWalletTxNavigator';
 import { backendNetworksActions, useBackendNetworksStore } from '@/features/network/stores/backendNetworksStore';
 import { ChainId } from '@/features/network/types/backendNetworks';
 import { trackSentTransaction } from '@/features/transfer/utils/trackSentTransaction';
@@ -35,7 +36,6 @@ import { greaterThan, lessThan } from '@/helpers/utilities';
 import * as i18n from '@/languages';
 import { ensureError, logger, RainbowError } from '@/logger';
 import { loadWallet, sendTransaction } from '@/model/wallet';
-import { setHardwareTXError } from '@/navigation/HardwareWalletTxNavigator';
 import { useNavigation } from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
 import { interactionsCountQueryKey } from '@/resources/addys/interactions';
