+++ src/navigation/types.ts
@@ -22,6 +22,7 @@ import { type RequestData, type RequestSource } from '@/features/dapp-request/ty
 import { type ENSProfile } from '@/features/ens/types/profile';
 import { type REGISTRATION_MODES } from '@/features/ens/utils/helpers';
 import { type GasTrend } from '@/features/gas/utils/meteorology';
+import { type LEDGER_ERROR_CODES } from '@/features/hardware-wallet/utils/ledger';
 import { type ChainId } from '@/features/network/types/backendNetworks';
 import { type HlTrade, type PerpMarket, type PerpsPosition, type TriggerOrderSource, type TriggerOrderType } from '@/features/perps/types';
 import { type PolymarketPosition } from '@/features/polymarket/types';
@@ -44,7 +45,6 @@ import { type Claimable, type RainbowClaimable } from '@/resources/addys/claimab
 import { type RevokeReason } from '@/screens/delegation/RevokeDelegationPanel';
 import { type ExpandedSheetParamAsset } from '@/screens/expandedAssetSheet/context/ExpandedAssetSheetContext';
 import { type AreAllKeysOptional, type ExtractOptionalKeys } from '@/types/objects';
-import { type LEDGER_ERROR_CODES } from '@/utils/ledger';
 
 export type PortalSheetProps = {
   children: React.FC;
